requirements.txt
arabic-reshaper==3.0.0
arrow==1.4.0
asgiref==3.11.0
asn1crypto==1.5.1
binaryornot==0.4.4
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
cryptography==46.0.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.8
django-allauth==0.57.2
django-allauth-2fa==0.12.0
django-axes==8.1.0
django-browser-reload==1.21.0
django-cors-headers==4.9.0
django-csp==4.0
django-daisy==2.0.7
django-formtools==2.5.1
django-guardian==3.2.0
django-otp==1.7.0
django-tailwind==4.4.2
django-unfold==0.76.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
et_xmlfile==2.0.0
fonttools==4.61.1
freetype-py==2.5.1
honcho==2.0.0
html5lib==1.1
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
numpy==2.3.5
oauthlib==3.3.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==2.23
pydyf==0.12.1
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
PyMySQL==1.1.2
pypdf==6.4.0
pyphen==0.17.2
pytailwindcss==0.3.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.1.0
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rlPyCairo==0.4.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.4
svglib==1.6.0
text-unidecode==1.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.5.0
weasyprint==67.0
webencodings==0.5.1
xhtml2pdf==0.2.17
xlrd==2.0.2
zopfli==0.4.0
